+++ src/DependencyInjection/Configuration.php
@@ -78,12 +78,15 @@ class Configuration implements ConfigurationInterface
                         ->end()
                             ->children()
                                 ->arrayNode('resolve_env')
+                                ->info('An environment variables resolving via consul')
                                 ->addDefaultsIfNotSet()
                                     ->children()
                                         ->booleanNode('enabled')
+                                            ->info('Resolving enabled')
                                             ->defaultTrue()
                                         ->end()
                                         ->scalarNode('cache')
+                                            ->info('TTL of entry if numeric or service identifier')
                                             ->defaultNull()
                                         ->end()
                                     ->end()
